+++ packages/react-components/src/button/index.js
@@ -1,7 +1,11 @@
 import PillButton from './components/pillButton'
+import IconButton from './components/iconButton'
+import IconWithTextButton from './components/iconWithTextButton'
 
-export { PillButton }
+export { PillButton, IconButton, IconWithTextButton }
 
 export default {
   PillButton,
+  IconButton,
+  IconWithTextButton,
 }