+++ packages/react-components/src/button/utils/theme.js
@@ -98,7 +98,142 @@ export const getOutlinePillButtonTheme = (theme, disabled) => {
   }
 }
 
+export const getPrimaryIconButtonTheme = (theme, isActive, isDisabled) => {
+  if (isDisabled) {
+    return {
+      color: colorGrayscale.gray400,
+      hoverColor: colorGrayscale.gray400,
+    }
+  }
+
+  const switchKey = isActive ? `${theme}-active` : theme
+  switch (switchKey) {
+    case 'photography':
+      return {
+        color: colorGrayscale.white,
+        hoverColor: colorSupportive.main,
+      }
+    case 'photography-active':
+      return {
+        color: colorSupportive.main,
+        hoverColor: colorSupportive.main,
+      }
+    case 'transparent':
+      return {
+        color: colorGrayscale.white,
+        hoverColor: colorGrayscale.gray200,
+      }
+    case 'transparent-active':
+      return {
+        color: colorGrayscale.white,
+        hoverColor: colorGrayscale.white,
+      }
+    case 'normal-active':
+      return {
+        color: colorBrand.heavy,
+        hoverColor: colorBrand.heavy,
+      }
+    case 'normal':
+    default:
+      return {
+        color: colorGrayscale.gray600,
+        hoverColor: colorGrayscale.gray800,
+      }
+  }
+}
+
+export const getSecondaryIconButtonTheme = (theme, isActive, isDisabled) => {
+  if (isDisabled) {
+    return {
+      color: colorGrayscale.gray400,
+      hoverColor: colorGrayscale.gray400,
+    }
+  }
+
+  const switchKey = isActive ? `${theme}-active` : theme
+  switch (switchKey) {
+    case 'photography':
+      return {
+        color: colorGrayscale.gray400,
+        hoverColor: colorSupportive.main,
+      }
+    case 'photography-active':
+      return {
+        color: colorSupportive.main,
+        hoverColor: colorSupportive.main,
+      }
+    case 'transparent':
+      return {
+        color: colorGrayscale.gray600,
+        hoverColor: colorGrayscale.white,
+      }
+    case 'transparent-active':
+      return {
+        color: colorGrayscale.gray600,
+        hoverColor: colorGrayscale.gray600,
+      }
+    case 'normal-active':
+      return {
+        color: colorBrand.heavy,
+        hoverColor: colorBrand.heavy,
+      }
+    case 'normal':
+    default:
+      return {
+        color: colorGrayscale.gray400,
+        hoverColor: colorGrayscale.gray600,
+      }
+  }
+}
+
+export const getIconWithTextButtonTheme = (theme, isActive, isDisabled) => {
+  if (isDisabled) {
+    return {
+      color: colorGrayscale.gray400,
+      hoverColor: colorGrayscale.gray400,
+    }
+  }
+
+  const switchKey = isActive ? `${theme}-active` : theme
+  switch (switchKey) {
+    case 'photography':
+      return {
+        color: colorGrayscale.gray100,
+        hoverColor: colorSupportive.main,
+      }
+    case 'photography-active':
+      return {
+        color: colorSupportive.main,
+        hoverColor: colorSupportive.main,
+      }
+    case 'transparent':
+      return {
+        color: colorGrayscale.gray100,
+        hoverColor: colorGrayscale.gray200,
+      }
+    case 'transparent-active':
+      return {
+        color: colorGrayscale.white,
+        hoverColor: colorGrayscale.white,
+      }
+    case 'normal-active':
+      return {
+        color: colorBrand.heavy,
+        hoverColor: colorBrand.heavy,
+      }
+    case 'normal':
+    default:
+      return {
+        color: colorGrayscale.gray600,
+        hoverColor: colorBrand.heavy,
+      }
+  }
+}
+
 export default {
   getFilledPillButtonTheme,
   getOutlinePillButtonTheme,
+  getPrimaryIconButtonTheme,
+  getSecondaryIconButtonTheme,
+  getIconWithTextButtonTheme,
 }